feat: adjusting package import and renaming email struct fields
